remove media_library install configs and add config to use them
the core.entity_form_mode.media.media_library.yml and
core.entity_view_mode.media.media_library.yml configs are in the
media_library module's `config/install` directory, so no need to include
those here, but for the media_library widget to work in selecting image
media especially the grid view, that view mode does need to be configured.

Not critical, as the form will still function without that config, but
it will look bad, so might as well include it as optional config here
alongside the new paragraph type that uses it.